testing/web-platform/tests/workers/Worker_ErrorEvent_message.htm

<!DOCTYPE html>
<html>
<head>
<title> AbstractWorker ErrorEvent.message </title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div id=log></div>
<script>
    
    var ErrorMessage = "Error Message";
    var FileName = './support/ErrorEvent.js';
    var description = "ErrorEvent message attribute represents the error message.";
    
    var t = async_test("Test Description: " + description);
                  
    var worker = new Worker(FileName);

    worker.onerror = t.step_func(function(evt)
    {
        assert_not_equals(evt.message, undefined);
        assert_not_equals(evt.message.indexOf('Error Message'), -1);
        t.done();        
    });
    
    worker.postMessage(ErrorMessage);
</script>
</body>
</html>
